fix(media): preserve direct-audio playback frontiers
Keep absolute PCM cursors at 64 bits, resume binding from the authoritative retired frontier, and exercise ISR retirement across DMA and counter wrap.

+++ ZZ9000_proto.sdk/ZZ9000OS/src/sdk_media_session.h
@@ -70,8 +70,8 @@ struct SDKMediaSessionAudioResult {
 struct SDKMediaAudioSource {
 	uint8_t *ring;
 	uint32_t capacity;
-	uint32_t produced_bytes;
-	uint32_t staged_bytes;
+	uint64_t produced_bytes;
+	uint64_t staged_bytes;
 	uint32_t sample_rate;
 	uint32_t channels;
 	uint32_t sample_format;
